libbinder: check null bytes in readString*Inplace
This is entirely defensive, since the only real guarantee we have here
from these APIs is that a buffer of a given length is available.
However, since we write 0's here, presumably to guard against people
assuming these are null-terminated strings, we might as well enforce
that they are actually null terminated.

Bug: 172655291
Test: binderParcelTest (added in newer CL)
Merged-In: Ie879112540155f6a93b97aeaf3d41ed8ba4ae79f
(cherry picked from commit 51e02b16c397c44ddf81a0736cf6045cd4c44128)

+++ libs/binder/Debug.cpp
@@ -208,7 +208,7 @@ void printHexData(int32_t indent, const void *buf, size_t length,
     }
 
     for (offset = 0; ; offset += bytesPerLine, pos += bytesPerLine) {
-        ssize_t remain = length;
+        long remain = length;
 
         char* c = buffer;
         if (!oneLine && !cStyle) {
